test(HeightTest): Height를 검증하는 테스트 메서드 위치를 LadderMakerTest -> HeightTest로 변경
